# Conflicts:
#	src/Zero.Web.Mvc/wwwroot/css/bootstrap.extend.min.css

+++ src/Zero.Core.Shared/Customize/NestedItem/NestedItem.cs
@@ -6,11 +6,11 @@ namespace Zero.Customize.NestedItem
 {
     public class  NestedItem
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string DisplayName { get; set; }
         
         [JsonProperty("parent_id")]
-        public long? ParentId { get; set; }
+        public int? ParentId { get; set; }
 
         [CanBeNull] public List<NestedItem> Children { get; set; }
     }
